ZEPPELIN-4747. Terminal interpreter support internal and external IP mapping
### What is this PR for?
When running the terminal interpreter in the notebook, the front end of the notebook needs to obtain the IP address of the server where the terminal interpreter is located to communicate.

In a public cloud environment, the cloud host has an internal IP and an external access IP, and the interpreter runs in the cloud host. This will cause the notebook front end to be unable to connect to the terminal interpreter properly, resulting in the terminal interpreter being unusable.

Solution: Set the mapping between internal IP and external IP in the terminal interpreter, and connect the front end of the notebook through the external IP of the terminal interpreter.

### What type of PR is it?
[Improvement]

### What is the Jira issue?
* https://jira.apache.org/jira/browse/ZEPPELIN-4747

### How should this be tested?
* First time? Setup Travis CI as described on https://zeppelin.apache.org/contribution/contributions.html#continuous-integration
* Strongly recommended: add automated unit tests for any new or changed behavior
* Outline any manual steps to test the PR here.

### Screenshots (if appropriate)

### Questions:
* Does the licenses files need update?
* Is there breaking changes for older versions?
* Does this needs documentation?

Author: Xun Liu <liuxun@apache.org>

Closes #3732 from liuxunorg/ZEPPELIN-4747 and squashes the following commits:

68f16caa4 [Xun Liu] Fixed
c462c3e46 [Xun Liu] ZEPPELIN-4747. Terminal interpreter support internal and external IP mapping