[ZEPPELIN-4776]. The same flink job url is duplicated in fronted for streaming job
### What is this PR for?

This PR is to remove duplicated flink job url. In ZEPPELIN-4759, I remove the refresh step in frontend, so it is not necessary to send job url again for each data refresh. So in this PR, I remove sendFlinkJobUrl in flink streaming job.

### What type of PR is it?
[Bug Fix ]

### Todos
* [ ] - Task

### What is the Jira issue?
* https://issues.apache.org/jira/browse/ZEPPELIN-4776

### How should this be tested?
* CI pass

### Screenshots (if appropriate)

### Questions:
* Does the licenses files need update? No
* Is there breaking changes for older versions? No
* Does this needs documentation? No

Author: Jeff Zhang <zjffdu@apache.org>

Closes #3749 from zjffdu/ZEPPELIN-4776 and squashes the following commits:

ff872efb9 [Jeff Zhang] [ZEPPELIN-4776]. The same flink job url is duplicated in fronted for streaming job