[ZEPPELIN-4763]. Use isolated interpreter for cron note job
### What is this PR for?

This PR is to use isolated interpreter for cron note job. Cron job is for production job, so we should use dedicated interpreter for it instead of using shared interpreter with other notes or users.  Currently interpreter binding mode is determined by note or users (scoped or isolated), This PR introduce `ExecutionContext` which include other factors (isCronMode) to determine how we create interpreters.

### What type of PR is it?
[Feature]

### Todos
* [ ] - Task

### What is the Jira issue?
* https://issues.apache.org/jira/browse/ZEPPELIN-4763

### How should this be tested?
* Unit test is updated, and manually tested

### Screenshots (if appropriate)

### Questions:
* Does the licenses files need update? No
* Is there breaking changes for older versions? No
* Does this needs documentation? No

Author: Jeff Zhang <zjffdu@apache.org>

Closes #3765 from zjffdu/ZEPPELIN-4763 and squashes the following commits:

39147c397 [Jeff Zhang] address comment
7bc690d95 [Jeff Zhang] [ZEPPELIN-4763]. Use isolated interpreter for cron note job