[ZEPPELIN-4870] Improve parsing of the paragraph properties
### What is this PR for?

We can provide properties that are local to the paragraph, that could be used to pass an additional information for interpreter that could affect its behavior.  Unfortunately existing parsing functionality relies on the fact that key/value pairs need to be separated by `,` character, and doesn't handle values with special characters (`,`, `=`, ...) inside, like this:

```
%cassandra(locale=ruRU, timeFormat="E, d MMM yy", floatPrecision = 5, outputFormat=cql)
```

This PR changes the parsing logic to perform character-by-character parsing, and handling of the quoted keys & values, escaping of the special characters, etc.

### What type of PR is it?

Improvement

### What is the Jira issue?
* https://issues.apache.org/jira/browse/ZEPPELIN-4870

### How should this be tested?
* https://travis-ci.org/github/alexott/zeppelin/builds/697522260
* additional unit tests were added

Author: Alex Ott <alexott@gmail.com>

Closes #3799 from alexott/ZEPPELIN-4870 and squashes the following commits:

5fb6ee84e [Alex Ott] [ZEPPELIN-4870] Improve parsing of the paragraph properties