[ZEPPELIN-4877]. text between 2 sql comment is mistaken as comment
### What is this PR for?

This PR fix the bug in Splitter that text between 2 sql comment is mistaken as comment. Unit test is added in this PR as well. Besides that, this PR also remove the comment in the middle of sql. e.g.
```
select a -- comment 1
from table_1
```
will be converted to
```
select a
from table_1
```

### What type of PR is it?
[Bug Fix]

### Todos
* [ ] - Task

### What is the Jira issue?
* https://issues.apache.org/jira/browse/ZEPPELIN-4877

### How should this be tested?
* CI pass

### Screenshots (if appropriate)

### Questions:
* Does the licenses files need update? No
* Is there breaking changes for older versions? No
* Does this needs documentation? No

Author: Jeff Zhang <zjffdu@apache.org>

Closes #3796 from zjffdu/ZEPPELIN-4877 and squashes the following commits:

0b0b72802 [Jeff Zhang] address comment
6e9287e16 [Jeff Zhang] address comment
63afd6ad3 [Jeff Zhang] [ZEPPELIN-4877]. text between 2 sql comment is mistaken as comment