[ZEPPELIN-4890]. Support to run note in isolated mode in rest api
### What is this PR for?

This PR is to support to run note in isolated mode in rest api, that means all the interpreter in this rest api is isolated and will be released after the note is finished. This is used for other third party that use rest api to integrate with zeppelin to run note, e.g. airflow use it to run ETL job.

### What type of PR is it?
[Improvement | Feature]

### Todos
* [ ] - Task

### What is the Jira issue?
* https://issues.apache.org/jira/browse/ZEPPELIN-4890

### How should this be tested?
* CI pass

### Screenshots (if appropriate)

### Questions:
* Does the licenses files need update? No
* Is there breaking changes for older versions? No
* Does this needs documentation? No

Author: Jeff Zhang <zjffdu@apache.org>

Closes #3821 from zjffdu/ZEPPELIN-4890 and squashes the following commits:

76f09955e [Jeff Zhang] [ZEPPELIN-4890]. Support to run note in isolated mode in rest api