[ZEPPELIN-4938] Allow execution of arbitrary CQL commands
### What is this PR for?

Right now, Cassandra interpreter is very restrictive & allows execution only of the predefined set of the commands, and this doesn't allow to execute CQL commands that are extension of the core of CQL.  This also may require additional changes in the Cassandra interpreter if language will be extended.  This PR allows to execute any command, not only hardcoded.

### What type of PR is it?
Improvement

### What is the Jira issue?
* https://issues.apache.org/jira/browse/ZEPPELIN-4938

### How should this be tested?
*  Tested manually
* https://travis-ci.org/github/alexott/zeppelin/jobs/705358648

Author: Alex Ott <alexott@apache.org>

Closes #3841 from alexott/ZEPPELIN-4938 and squashes the following commits:

91f003d4a [Alex Ott] [ZEPPELIN-4938] Allow execution of arbitrary CQL commands