[ZEPPELIN-4951]. Set paragraph status to be abort when recovery is disabled
### What is this PR for?

A simple PR which set paragraph status to be abort when loading note and recovery is disabled. Otherwise we would see paragraph is in running state and we are unable to abort it because we have lost the connection to the interpreter process when recovery is disabled.

### What type of PR is it?
[Bug Fix ]

### Todos
* [ ] - Task

### What is the Jira issue?
* https://issues.apache.org/jira/browse/ZEPPELIN-4951

### How should this be tested?
* CI

### Screenshots (if appropriate)

### Questions:
* Does the licenses files need update? NO
* Is there breaking changes for older versions? No
* Does this needs documentation? No

Author: Jeff Zhang <zjffdu@apache.org>

Closes #3853 from zjffdu/ZEPPELIN-4951 and squashes the following commits:

e5d211367 [Jeff Zhang] [ZEPPELIN-4951]. Set paragraph status to be abort when recovery is disabled