[ZEPPELIN-4970]. Pass weburl from interpreter process to InterpreterGroup of zeppelin server
### What is this PR for?

The purpose of this PR is to pass weburl from interpreter process to interpreter group of zeppelin server, so that we can display the weburl on zeppelin web page or pass the weburl to users via rest api. The weburl is pretty important for user to manage their interpreter process. In this PR, I pass spark and flink interpreter's weburl to zeppelin server.

### What type of PR is it?
[ Feature ]

### Todos
* [ ] - Task

### What is the Jira issue?
* https://issues.apache.org/jira/browse/ZEPPELIN-4970

### How should this be tested?
* CI pass

### Screenshots (if appropriate)

### Questions:
* Does the licenses files need update? No
* Is there breaking changes for older versions? No
* Does this needs documentation? No

Author: Jeff Zhang <zjffdu@apache.org>

Closes #3866 from zjffdu/ZEPPELIN-4970 and squashes the following commits:

316c34d54 [Jeff Zhang] [ZEPPELIN-4970]. Pass weburl from interpreter process to InterpreterGroup of zeppelin server