[ZEPPELIN-5055]. Interpreter log is missing in zeppelin server docker container
### What is this PR for?

The root cause in the log4j.properties in the zeppelin server image. For k8s, we should use stdout for logging, but for the single zeppelin server container, we should use dailyfile for logging (both zeppelin server and interpreter). This PR add a new env `ZEPPELIN_IN_DOCKER` to choose the right log4j properties file for the right scenario.

### What type of PR is it?
[Bug Fix ]

### Todos
* [ ] - Task

### What is the Jira issue?
* https://issues.apache.org/jira/browse/ZEPPELIN-5055

### How should this be tested?
* CI pass

### Screenshots (if appropriate)

### Questions:
* Does the licenses files need update? No
* Is there breaking changes for older versions? No
* Does this needs documentation? No

Author: Jeff Zhang <zjffdu@apache.org>

Closes #3918 from zjffdu/ZEPPELIN-5055 and squashes the following commits:

8858569a0 [Jeff Zhang] [ZEPPELIN-5055]. Interpreter log is missing in zeppelin server docker container