[ZEPPELIN-5060] Disabled paragraph cause Zeppelin notebook to hang
### What is this PR for?

When trying to run a disabled paragraph, backend hangs because the paragraph is not started, but we are still waiting for it to finish. Instead, we should skip it just like empty paragraphs.

Also, we need to change paragraph state to pending, so next state change to finished is broadcasted to client.

### What type of PR is it?

Bug Fix

### Todos

### What is the Jira issue?

https://issues.apache.org/jira/browse/ZEPPELIN-5060

### How should this be tested?

### Screenshots (if appropriate)

### Questions:

Author: Timo Olkkonen <timo.p.olkkonen@gmail.com>

Closes #3932 from olkkoti/ZEPPELIN-5060 and squashes the following commits:

9de27f072 [Timo Olkkonen] Update integration test: disabled paragraph should be FINISHED after run all.
921e3dfdb [Timo Olkkonen] Set Paragraph status to Pending at start of execution, and skip if it is disabled.