[Hotfix] Add logging libraries into zeppelin interpreter process
### What is this PR for?
With #3914 I removed the logging libraries for Zeppelin-Interpreters which are not shipped with their own logging libraries (e.g. Markdown). The logging libraries are not part of zeppelin-interpreter-shaded. With this change we add the logging libraries back to interpreter-parent, so that they are present in every Zeppelin-Interpreters process.

### What type of PR is it?
 - Hot Fix

### How should this be tested?
* Travis-CI: https://travis-ci.org/github/Reamer/zeppelin/builds/732966733

### Questions:
* Does the licenses files need update? No
* Is there breaking changes for older versions? No
* Does this needs documentation? No

Author: Philipp Dallig <philipp.dallig@gmail.com>

Closes #3934 from Reamer/hotfix_logging and squashes the following commits:

39b8edbcd [Philipp Dallig] Add logging libraries into zeppelin interpreter process

(cherry picked from commit 82a0bae78eb52649536468136174d8c5f2fc3a75)
Signed-off-by: Philipp Dallig <philipp.dallig@gmail.com>