ZEPPELIN-5081: Add loopPostInit to bind $intp
### What is this PR for?
There is a bug in the project: $intp variable is not defined for Spark/Scala 2.12 interpreter. It looks like this part of REPL in Spark 2.12 and in Spark 2.11 is the same, so I use the code from Spark/Scala 2.11 interpreter to fix it.

### What type of PR is it?
Bug Fix

### What is the Jira issue?
[ZEPPELIN-5081]

### How should this be tested?
* Tested manually with master and Spark 2.4.2/Scala 2.12.8

Author: khud <vitaly.khudobakhshov@gmail.com>

Closes #3935 from khud/fix_missing_intp and squashes the following commits:

301b7a440 [khud] Add test for $intp
142cf6f55 [khud] Add loopPostInit to bind $intp

(cherry picked from commit fbf3019efdd7381ca7e77b994d0a94f6cb8ea383)
Signed-off-by: Jeff Zhang <zjffdu@apache.org>