[ZEPPELIN-5112]. upgrade-note.sh throws exception while upgrading notes on hdfs storage
### What is this PR for?

The root cause is that we didn't put hadoop jars in the classpath in `upgrade-note.sh`. This PR update script `upgrade-note.sh`. Besides that this PR also fix an upgrading issue in `ParagraphRuntimeInfo.java`. The `values` is List<String> in old version of zeppelin, but it is changed to List<Map<String, String>> in new version of zeppelin.

### What type of PR is it?
[Bug Fix]

### Todos
* [ ] - Task

### What is the Jira issue?
* https://issues.apache.org/jira/browse/ZEPPELIN-5112

### How should this be tested?
* CI pass
https://travis-ci.org/github/zjffdu/zeppelin/builds/739821895

### Screenshots (if appropriate)

### Questions:
* Does the licenses files need update? No
* Is there breaking changes for older versions? No
* Does this needs documentation? No

Author: Jeff Zhang <zjffdu@apache.org>

Closes #3958 from zjffdu/ZEPPELIN-5112 and squashes the following commits:

5e74dae1c [Jeff Zhang] [ZEPPELIN-5112]. upgrade-note.sh throws exception while upgrading notes on hdfs storage