[ZEPPELIN-5147]. ConcurrentModificationException in Note#toJson
### What is this PR for?

The root cause is that when calling Note#toJson, note may be in change (adding/removing paragraph). In this PR, I change Note#paragraphs to CopyOnWriteArrayList, because adding/removing operation for paragraph is not a frequent operation, so it is ok for use CopyOnWriteArrayList here.
Besides that, this PR also do some improvement on other parts, such as adding more logging.

### What type of PR is it?
[Bug Fix | Improvement ]

### Todos
* [ ] - Task

### What is the Jira issue?
* https://issues.apache.org/jira/browse/ZEPPELIN-5147

### How should this be tested?
* CI pass

### Screenshots (if appropriate)

### Questions:
* Does the licenses files need update? No
* Is there breaking changes for older versions? No
* Does this needs documentation? No

Author: Jeff Zhang <zjffdu@apache.org>

Closes #3985 from zjffdu/ZEPPELIN-5147 and squashes the following commits:

1bc9bbccd [Jeff Zhang] [ZEPPELIN-5147]. ConcurrentModificationException in Note#toJson