[ZEPPELIN-5152]. Paragraph is not released after it is removed
### What is this PR for?

The root cause is in InterpreterContext#build, each time it would put its current thread into its static map, but never released. This make the thread is never garbage collected. This PR fix this issue, besides that it also fix minor issues where user paragraph should not created when personalize is not enabled.

### What type of PR is it?
[Bug Fix ]

### Todos
* [ ] - Task

### What is the Jira issue?
* https://issues.apache.org/jira/projects/ZEPPELIN/issues/ZEPPELIN-5152?filter=allopenissues

### How should this be tested?
* CI pass

### Screenshots (if appropriate)

### Questions:
* Does the licenses files need update? No
* Is there breaking changes for older versions? No
* Does this needs documentation? No

Author: Jeff Zhang <zjffdu@apache.org>

Closes #3987 from zjffdu/ZEPPELIN-5152 and squashes the following commits:

03c5902b1 [Jeff Zhang] [ZEPPELIN-5152]. Paragraph is not released after it is removed