[ZEPPELIN-5213]. Move spark-submit as a separated interpreter module
### What is this PR for?

This PR is to move spark-submit from spark module to another separated interpreter module. Because it doesn't make sense to run spark-submit in the same JVM of spark interpreter. It would cause some very confused things. e.g. If running it in yarn mode, the yarn app would never be registered unless SparkContext is created. But if you only `%spark.submit`, SparkContext would never be created.

### What type of PR is it?
[ Refactoring]

### Todos
* [ ] - Task

### What is the Jira issue?
* https://issues.apache.org/jira/browse/ZEPPELIN-5213

### How should this be tested?
* Manually tested

### Screenshots (if appropriate)

### Questions:
* Does the licenses files need update? No
* Is there breaking changes for older versions? No
* Does this needs documentation? No

Author: Jeff Zhang <zjffdu@apache.org>

Closes #4032 from zjffdu/ZEPPELIN-5213 and squashes the following commits:

382755877 [Jeff Zhang] [ZEPPELIN-5213]. Move spark-submit as a separated interpreter module