[ZEPPELIN-5192]: Don't try to decrypt a non-existent credentials file
### What is this PR for?

This is a very simple fix for the issue described in [ZEPPELIN-5192](https://issues.apache.org/jira/browse/ZEPPELIN-5192): Zeppelin will currently fail to start if a credentials encryption key is configured but no `credentials.json` file exists currently. This is due to an NPE inside the decryption code path. The fix is to only attempt the decryption if the file could be loaded.

### What type of PR is it?

[Bug Fix]

### What is the Jira issue?

* [ZEPPELIN-5192](https://issues.apache.org/jira/browse/ZEPPELIN-5192)

### How should this be tested?

Booting a fresh Zeppelin server with ZEPPELIN_CREDENTIALS_ENCRYPT_KEY set is sufficient to reproduce the issue and validate the fix.

Author: Niels Grewe <niels@zasta.de>

Closes #4062 from ngrewe/bugfix/ZEPPELIN-5192 and squashes the following commits:

6dd974358 [Niels Grewe] fix(ZEPPELIN-5192): Don't try to decrypt a non-existent credentials file