InterpreterSetting: don't crash when sorting settings.
The existing comparator would return -1 for both (a, b) and (b, a),
if both of these settings keys are not present in the template,
and such behaviour violates required comparator properties.

The issue manifests with exception in the form

  java.lang.IllegalArgumentException: Comparison method violates its general contract!
  at java.util.TimSort.mergeLo(TimSort.java:777)

and prevents server from starting up.

### What type of PR is it?
Bug Fix

### Questions:
* Does the licenses files need update? I don't think, the change is trivial.
* Is there breaking changes for older versions? No.
* Does this needs documentation? No.

Author: Vladimir Prus <vladimir.prus@gmail.com>

Closes #4075 from vprus/interpreter-settings-sorting-v2 and squashes the following commits:

7cd0b6485 [Vladimir Prus] Cosmetics
47e129a31 [Vladimir Prus] Remove spurious change
0cbc70fbd [Vladimir Prus] Cosmetrics.
e032d9990 [Vladimir Prus] InterpreterSetting: don't crash when sorting settings.