[ZEPPELIN-5439] Improve the logic of extract hive job url in JdbcInterpreter
### What is this PR for?

Previously we extract hive job url from hive jdbc job log, but it seems the in the latest hive, there's no job url info in log.
This PR is use another approach to do that, we use the paragraphId as the hive job yarn tag, after hive sql execution is started, we look at the yarn app that has this tag.

### What type of PR is it?
[Improvement]

### Todos
* [ ] - Task

### What is the Jira issue?
* https://issues.apache.org/jira/browse/ZEPPELIN-5439

### How should this be tested?
* CI pass

### Screenshots (if appropriate)

### Questions:
* Does the licenses files need update? No
* Is there breaking changes for older versions? No
* Does this needs documentation? No

Author: Jeff Zhang <zjffdu@apache.org>

Closes #4166 from zjffdu/ZEPPELIN-5439 and squashes the following commits:

c88c528299 [Jeff Zhang] [ZEPPELIN-5439] Improve the logic of extract hive job url in JdbcInterpreter