[ZEPPELIN-6054] Eliminate no-invalid-this warnings
### What is this PR for?
This PR aims to refactor several parts of the Zeppelin project codebase to eliminate no-invalid-this warnings reported by ESLint. These warnings indicate that the this keyword is being used in ways that might lead to unexpected behavior or bugs, particularly within nested callbacks or when this loses its intended context. By addressing these warnings, we improve the stability, readability, and maintainability of the codebase.

**Note on `note-action.service.js`**:
Initially, I aimed to refactor `note-action.service.js` to eliminate ESLint warnings. However, during testing, it was found that the note actions on the default page were not displaying correctly after the changes. To ensure the stability of the project, I have reverted the changes made to `note-action.service.js`. I plan to revisit this refactor to address the ESLint warnings while ensuring no issues arise in the build process.

### What type of PR is it?
Refactoring

### Todos
* [x] - Resolve ESLint warnings of files in zeppelin-web/src/app
* [ ] - Resolve ESLint warnings of files in zeppelin-web/src/components
* [x] - Check if project compiles successfully

### What is the Jira issue?
[ZEPPELIN-6054](https://issues.apache.org/jira/browse/ZEPPELIN-6054)

### How should this be tested?
* CI
* Build and run 

### Screenshots (if appropriate)
*Current Warnings*:
<img width="1102" alt="image" src="https://github.com/user-attachments/assets/4a928656-ec97-4553-a12c-2cd0fbd46350">


### Questions:
* Does the license files need to update? - No
* Is there breaking changes for older versions? - No
* Does this needs documentation? - No


Closes #4788 from cherrie-k/fix/minor-warnings.

Signed-off-by: Cheng Pan <chengpan@apache.org>