[ZEPPELIN-6111] Update to include optional classic UI instructions
### What is this PR for?
We've recently switched the default UI to the new version.

As a result, it's important to provide instructions for users who want to access the previous default UI.

This is a minimal guide for now, and more detailed documentation should be added in the future.

### What type of PR is it?
Documentation

### Todos
* [ ] - Task

### What is the Jira issue?
* Open an issue on Jira https://issues.apache.org/jira/browse/ZEPPELIN/6111

### How should this be tested?
* Strongly recommended: add automated unit tests for any new or changed behavior
* Outline any manual steps to test the PR here.

### Screenshots (if appropriate)

### Questions:
* Does the license files need to update? No
* Is there breaking changes for older versions? No
* Does this needs documentation? No

Closes #4847 from tbonelee/add-guidance-to-classic-ui.

Signed-off-by: Cheng Pan <chengpan@apache.org>
(cherry picked from commit 4aff03bb711781242e6e0d4f22d6c0fc16be65eb)
Signed-off-by: Cheng Pan <chengpan@apache.org>