[ZEPPELIN-3317] Update personalized mode document page
> What is this PR for?

This PR is to populate the "Personalized Mode" portion of the project that is previously devoid of any information whatsoever.
> What type of PR is it?

This would be initial Documentation for this repo

> What is the Jira issue?

>    Open an issue on Jira https://issues.apache.org/jira/browse/ZEPPELIN/

[ZEPPELIN-3317](https://issues.apache.org/jira/browse/ZEPPELIN-3317)

> How should this be tested?

No testing needed.

>    Strongly recommended: add automated unit tests for any new or changed behavior
Not needed- this is initial documentation

>     Outline any manual steps to test the PR here.

There are not steps- this is initial documentation.

> Screenshots (if appropriate)

N/A

Questions:

>     Does the license files need to update?
No license update needed.

>     Is there breaking changes for older versions?
All previous versions are blank. No breaking would occur.
>     Does this needs documentation?
This is the said documentation

Closes #4821 from chuckn408/master.

Signed-off-by: Cheng Pan <chengpan@apache.org>
(cherry picked from commit c5ccc8e584934fa4f96de486cd97312180147b1f)
Signed-off-by: Cheng Pan <chengpan@apache.org>